+++ scripts/generate-apps.cjs
@@ -11,7 +11,7 @@ const sharp = require('sharp');
 // ---------------------------------------------------------
 const OUTPUT_PATH = path.resolve('./public/data/apps-data.json');
 const LOGO_DIR = path.resolve('./public/assets/apps');
-const SOURCE_URL = 'https://raw.githubusercontent.com/DesignLipsx/WinUI-3-Apps-List/refs/heads/main/README.md';
+const SOURCE_URL = 'https://raw.githubusercontent.com/jishnu-kv/WinUI-3-Apps-List/refs/heads/main/README.md';
 
 
 // ---------------------------------------------------------